requirements.txt
pymongo
groq
requests
python-dotenv
streamlit
Pillow  # For logo handling